Remove upload and release functions
We are going to these in CI workflows instead.

This change greatly simplifies the project.
