--- lib/get-major.ts
@@ -1,4 +1,0 @@
-export function getMajor(nodeVersion: string) {
-  const [, version] = nodeVersion.match(/^v?(\d+)/) || ['', 0];
-  return Number(version) | 0;
-}
