--- lib/temp-path.ts
@@ -1,5 +1,0 @@
-import uniqueTempDir from 'unique-temp-dir';
-
-export function tempPath(...args: Parameters<typeof uniqueTempDir>) {
-  return uniqueTempDir(...args);
-}
